Add admin check in end_command_handler

+++ services/handler/app/handlers/chats/end.py
@@ -10,6 +10,9 @@ from app.utils.end_message import get_end_message
 async def end_command_handler(bot, message: MessageSchema, sql_session=None) -> None:
     session = await bot.app.store.sessions.get_session_by_chat_id(sql_session=sql_session, chat_id=message.chat.id)
 
+    if not bot.is_user_admin(chat_id=message.chat.id, user_id=message.message_from.id):
+        raise RuntimeError
+
     await delete_session(store=bot.app.store,
                          sql_session=sql_session,
                          session=session)
